UI Shop now opens clearly, with all details.
 Stil need to figure out a way to take 'cards and place them in your own inventory.

+++ Assets/Scripts/Ingredient.cs
@@ -8,7 +8,7 @@ public class Ingredient : MonoBehaviour {
 	public Sprite icon;
 	public string title;
 	public string description;
-	public float freshness;
+	public int freshness;
 
 	void Start () {
